Move messages to and from worker into the worker module
